Добавлено использование интерфейсов в сервисах.
